requirements.txt
langchain
streamlit
easyocr
numpy
pillow
serpapi